# server/openApi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/login": {
      "post": {
        "operationId": "MainController_login",
        "summary": "用户登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ]
      }
    },
    "/logout": {
      "post": {
        "operationId": "MainController_logout",
        "summary": "退出登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ]
      }
    },
    "/register": {
      "post": {
        "operationId": "MainController_register",
        "summary": "用户注册",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ]
      }
    },
...
    "/getTenantList": {
      "get": {
        "operationId": "MainController_getTenantList",
        "summary": "获取全部租户",
        "parameters": [
          {
            "name": "tenantName",
            "required": false,
            "in": "query",
            "description": "租户名称",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ]
      }
    },
    "/system/tenant": {
      "post": {
        "operationId": "SysTenantController_create",
        "summary": "租户管理-创建",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSysTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "租户管理"
        ]
      },
      "put": {
        "operationId": "SysTenantController_update",
        "summary": "租户管理-修改",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSysTenantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "租户管理"
        ]
      }
    },
    "/system/tenant/list": {
      "get": {
        "operationId": "SysTenantController_findAll",
        "summary": "租户管理-列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantCode",
            "required": false,
            "in": "query",
            "description": "租户编码",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "tenantName",
            "required": false,
            "in": "query",
            "description": "租户名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "租户状态",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantContact",
            "required": false,
            "in": "query",
            "description": "联系人",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantPhone",
            "required": false,
            "in": "query",
            "description": "联系电话",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "租户管理"
        ]
      }
    },
    "/system/tenant/{id}": {
      "get": {
        "operationId": "SysTenantController_findOne",
        "summary": "租户管理-详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "租户管理"
        ]
      },
      "delete": {
        "operationId": "SysTenantController_remove",
        "summary": "租户管理-删除",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "租户管理"
        ]
      }
    },
...
    "/system/tenant/package": {
      "post": {
        "operationId": "SysTenantPackageController_create",
        "summary": "租户套餐-创建",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePackageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "租户套餐"
        ]
      },
      "put": {
        "operationId": "SysTenantPackageController_update",
        "summary": "租户套餐-修改",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePackageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "租户套餐"
        ]
      }
    },
    "/system/tenant/package/list": {
      "get": {
        "operationId": "SysTenantPackageController_findAll",
        "summary": "租户套餐-列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "packageName",
            "required": false,
            "in": "query",
            "description": "套餐名称",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "租户套餐"
        ]
      }
    },
    "/system/tenant/package/{id}": {
      "get": {
        "operationId": "SysTenantPackageController_findOne",
        "summary": "租户套餐-详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "租户套餐"
        ]
      },
      "delete": {
        "operationId": "SysTenantPackageController_remove",
        "summary": "租户套餐-删除",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "租户套餐"
        ]
      }
    }
  },
  "info": {
    "title": "Nest-Admin",
    "description": "Nest-Admin 接口文档",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "tenantCode": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password",
          "uuid"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "tenantCode": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password",
          "uuid"
        ]
      },
      "CreateSysTenantDto": {
        "type": "object",
        "properties": {
          "tenantName": {
            "type": "string",
            "description": "租户名称"
          },
          "userName": {
            "type": "string",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "description": "密码"
          },
          "tenantDesc": {
            "type": "string",
            "description": "租户描述"
          },
          "tenantContact": {
            "type": "string",
            "description": "租户联系人"
          },
          "tenantPhone": {
            "type": "string",
            "description": "租户联系电话"
          },
          "tenantArea": {
            "type": "string",
            "description": "租户地区"
          },
          "tenantAddress": {
            "type": "string",
            "description": "租户联系地址"
          },
          "longitude": {
            "type": "number",
            "description": "经度"
          },
          "latitude": {
            "type": "number",
            "description": "纬度"
          },
          "tenantLogo": {
            "type": "string",
            "description": "租户logo"
          },
          "expireTime": {
            "format": "date-time",
            "type": "string",
            "description": "过期时间"
          },
          "packageId": {
            "type": "number",
            "description": "租户套餐id"
          },
          "areaCode": {
            "type": "string",
            "description": "租户地区编码"
          }
        },
        "required": [
          "tenantName",
          "userName",
          "password",
          "tenantArea",
          "tenantAddress",
          "longitude",
          "latitude",
          "tenantLogo",
          "expireTime",
          "packageId",
          "areaCode"
        ]
      },
      "UpdateSysTenantDto": {
        "type": "object",
        "properties": {
          "tenantName": {
            "type": "string",
            "description": "租户名称"
          },
          "userName": {
            "type": "string",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "description": "密码"
          },
          "tenantDesc": {
            "type": "string",
            "description": "租户描述"
          },
          "tenantContact": {
            "type": "string",
            "description": "租户联系人"
          },
          "tenantPhone": {
            "type": "string",
            "description": "租户联系电话"
          },
          "tenantArea": {
            "type": "string",
            "description": "租户地区"
          },
          "tenantAddress": {
            "type": "string",
            "description": "租户联系地址"
          },
          "longitude": {
            "type": "number",
            "description": "经度"
          },
          "latitude": {
            "type": "number",
            "description": "纬度"
          },
          "tenantLogo": {
            "type": "string",
            "description": "租户logo"
          },
          "expireTime": {
            "format": "date-time",
            "type": "string",
            "description": "过期时间"
          },
          "packageId": {
            "type": "number",
            "description": "租户套餐id"
          },
          "areaCode": {
            "type": "string",
            "description": "租户地区编码"
          },
          "tenantId": {
            "type": "number",
            "description": "租户id"
          }
        },
        "required": [
          "tenantName",
          "userName",
          "password",
          "tenantArea",
          "tenantAddress",
          "longitude",
          "latitude",
          "tenantLogo",
          "expireTime",
          "packageId",
          "areaCode",
          "tenantId"
        ]
      },
...
      "CreatePackageDto": {
        "type": "object",
        "properties": {
          "packageName": {
            "type": "string",
            "description": "套餐名称"
          },
          "menuIds": {
            "type": "string",
            "description": "关联菜单id"
          },
          "menuCheckStrictly": {
            "type": "number",
            "description": "菜单树选择项是否关联显示"
          }
        }
      },
      "UpdatePackageDto": {
        "type": "object",
        "properties": {
          "packageName": {
            "type": "string",
            "description": "套餐名称"
          },
          "menuIds": {
            "type": "string",
            "description": "关联菜单id"
          },
          "menuCheckStrictly": {
            "type": "number",
            "description": "菜单树选择项是否关联显示"
          },
          "packageId": {
            "type": "number",
            "description": "套餐id"
          }
        }
      }
    }
  }
}